docs: μεταφορές και αναδιοργάνωση αρχείων τεκμηρίωσης και python_client (moved/renamed docs)
