--- EncoderReader/MD/README_HighPerformance.md
@@ -1,137 +1,0 @@
-# High-Performance Encoder Implementation
-
-## Overview
-This is an optimized version of the ESP32-S3 quadrature encoder reader, designed for maximum performance and reliability in demanding applications.
-
-## Key Improvements
-
-### 1. **Hardware PCNT Support (Primary Mode)**
-- **What:** Uses ESP32's dedicated Pulse Counter (PCNT) peripheral instead of software ISR
-- **Why:** Eliminates ISR overhead, prevents missed pulses at high speeds
-- **Performance:** Can handle >100kHz quadrature signals without CPU load
-- **Configuration:** Set `USE_HARDWARE_PCNT = 1` in config.h
-
-### 2. **Optimized ISR Mode (Fallback)**
-- **Direct GPIO Access:** Uses register reads instead of digitalRead() for 10x speed improvement
-- **Glitch Filtering:** Ignores edges closer than `MIN_EDGE_INTERVAL_US` microseconds
-- **Minimal ISR Code:** Ultra-short interrupt handlers to minimize latency
-
-### 3. **Signed Edge-Based Velocity**
-- **Problem:** Original edge velocity calculation lost direction information
-- **Solution:** Tracks sign of last delta and applies it to edge-based speed calculation
-- **Result:** Accurate velocity readings in both directions at all speeds
-
-### 4. **Adaptive Velocity Blending**
-- **Low Speed (<10 cps):** Uses window-based calculation (more stable)
-- **High Speed (>1000 cps):** Prefers edge-based calculation (more responsive)  
-- **Medium Speed:** Balanced 50/50 blend
-- **Benefits:** Best of both worlds - stability + responsiveness
-
-### 5. **Velocity Timeout**
-- **Problem:** EMA filter took too long to decay to zero when encoder stopped
-- **Solution:** Force velocity to zero if no edges detected for `VELOCITY_TIMEOUT_US`
-- **Result:** Immediate zero reading when encoder stops
-
-### 6. **Enhanced Configuration**
-```cpp
-#define USE_HARDWARE_PCNT  1      // Enable hardware counter
-#define MIN_EDGE_INTERVAL_US 10   // Glitch filter (microseconds)  
-#define VELOCITY_TIMEOUT_US 500000 // Zero velocity timeout (500ms)
-#define ADAPTIVE_BLENDING 1       // Smart velocity calculation
-```
-
-## Performance Comparison
-
-| Feature | Original | High-Performance | Improvement |
-|---------|----------|------------------|-------------|
-| Max Encoder Speed | ~50kHz | >100kHz | 2x+ |
-| CPU Usage @ 50kHz | ~15% | <2% | 7x better |
-| Velocity Accuracy | Good | Excellent | Direction preserved |
-| Glitch Immunity | Basic | Advanced | Hardware + software filtering |
-| Response Time | 50ms | 10ms | 5x faster |
-| Stop Detection | ~2 seconds | 0.5 seconds | 4x faster |
-
-## Hardware Requirements
-
-### Encoder Connection
-- **A Channel:** GPIO16 (black wire)
-- **B Channel:** GPIO17 (white wire)  
-- **Z Index:** GPIO18 (orange wire, optional)
-- **Power:** 5V to encoder, 3.3V pull-ups on signals
-
-### Signal Conditioning
-For maximum performance at high speeds:
-- Use 4.7kΩ pull-ups to 3.3V on A/B/Z lines
-- Consider 100pF capacitors to ground for noise filtering
-- Keep encoder cables short (<2m) and shielded
-- Separate encoder power from motor power
-
-## Configuration Options
-
-### PCNT vs ISR Mode
-```cpp
-#define USE_HARDWARE_PCNT 1  // Recommended for high performance
-```
-- **PCNT Mode:** Hardware counting, minimal CPU usage, handles >100kHz
-- **ISR Mode:** Software counting, more CPU usage, ~50kHz max
-
-### Velocity Calculation
-```cpp
-#define ADAPTIVE_BLENDING 1  // Recommended
-```
-- **Adaptive:** Automatically chooses best method based on speed
-- **Fixed:** Always uses 50/50 blend of window and edge calculations
-
-### Filtering
-```cpp
-#define MIN_EDGE_INTERVAL_US 10    // Adjust based on max expected speed
-#define VELOCITY_TIMEOUT_US 500000 // Adjust based on application needs
-```
-
-## New Commands
-- **ZERO:** Reset encoder position to zero (in addition to existing TARE, CAL, etc.)
-
-## Monitoring Performance
-
-The system reports its configuration at startup:
-```
-ESP32-S3 High-Performance Quadrature Encoder
-PPR=1024, Sample Rate=10ms
-Mode: Hardware PCNT (Maximum Performance)
-Velocity: Adaptive Window/Edge Blending
-Glitch Filter: 10 microseconds
-Velocity Timeout: 500 ms
-```
-
-## Troubleshooting
-
-### High Speed Issues
-- Reduce `MIN_EDGE_INTERVAL_US` if losing counts at high speed
-- Ensure power supply can handle encoder current
-- Check signal integrity with oscilloscope
-
-### Noise/Glitches
-- Increase `MIN_EDGE_INTERVAL_US` 
-- Add hardware filtering (RC + Schmitt trigger)
-- Use shielded cables
-
-### Velocity Instability  
-- Adjust `EMA_ALPHA` (higher = more responsive, lower = smoother)
-- Tune `VELOCITY_TIMEOUT_US` for your application
-- Consider mechanical coupling issues
-
-## Technical Details
-
-### PCNT Configuration
-- **Mode:** Quadrature decoding with A/B channels
-- **Counter:** 16-bit with overflow handling
-- **Filter:** 1µs hardware glitch rejection  
-- **Interrupts:** Overflow/underflow for extended range
-
-### ISR Optimization
-- **GPIO Access:** Direct register reads (`GPIO.in`)
-- **State Machine:** Optimized quadrature decode table
-- **Atomic Operations:** Careful interrupt disable/enable
-- **Memory:** IRAM placement for zero-latency execution
-
-This implementation provides industrial-grade performance suitable for high-speed automation, robotics, and precision measurement applications.
