+++ python_client/MD/COMPARISON.md
