+++ python_client/encoder_gui_modular.py
@@ -1,318 +1,250 @@
-#!/usr/bin/env python3
 """
-ESP32 Encoder Monitor GUI - Modular Version
-
-A professional GUI application for monitoring ESP32 quadrature encoder
-and load cell data with real-time plotting and data export capabilities.
-
-This modular version splits the original monolithic code into focused
-components for better maintainability and understanding.
+Main GUI application for ESP32 encoder monitoring.
+Modular design with separate components for data handling, visualization, and communication.
 """
-
 import threading
 import time
 import tkinter as tk
+from tkinter import ttk, messagebox, filedialog
+import pandas as pd
 from typing import Optional
 
-# Import our modular components
-from config import *
-from data_models import DataBuffer, Sample
-from serial_handler import SerialManager, get_available_ports
-from data_parser import DataParser
-from gui_components import MainWindow, DialogHelper
-from data_export import DataExporter
+from data_models import DataBuffer
+from serial_handler import SerialThread, get_available_ports, find_esp32_port
+from visualization import EncoderPlot, DataTable
 
 
-class EncoderApplication:
-    """Main application class that coordinates all components."""
+class EncoderGUI:
+    """Main GUI application for encoder monitoring."""
     
-    def __init__(self, root: tk.Tk):
-        self.root = root
+    def __init__(self):
+        # Initialize main window
+        self.root = tk.Tk()
+        self.root.title("ESP32 Encoder Monitor")
+        self.root.geometry("1000x700")
         
-        # Initialize state variables
-        self.selected_port = tk.StringVar()
+        # Data management
+        self.buffer = DataBuffer()
+        self.mutex = threading.Lock()
+        
+        # Communication
+        self.serial_thread: Optional[SerialThread] = None
         self.connection_state = tk.BooleanVar(value=False)
         self.running = tk.BooleanVar(value=False)
-        self.autoscroll = tk.BooleanVar(value=True)
         
-        # Initialize components
-        self.data_buffer = DataBuffer()
-        self.serial_manager = SerialManager()
-        self.data_parser = DataParser()
-        self.main_window = MainWindow(root)
+        # Create GUI components
+        self._create_widgets()
+        self._setup_periodic_tasks()
         
-        # Thread synchronization
-        self.data_mutex = threading.Lock()
+        # Try to auto-detect ESP32 port
+        auto_port = find_esp32_port()
+        if auto_port:
+            self.port_var.set(auto_port)
+    
+    def _create_widgets(self):
+        """Create all GUI widgets."""
+        # Main container
+        main_frame = ttk.Frame(self.root, padding="10")
+        main_frame.grid(row=0, column=0, sticky="nsew")
+        self.root.columnconfigure(0, weight=1)
+        self.root.rowconfigure(0, weight=1)
         
-        # Initialize UI and start periodic tasks
-        self._setup_ui()
-        self._start_periodic_tasks()
+        # Connection frame
+        conn_frame = ttk.LabelFrame(main_frame, text="Connection", padding="5")
+        conn_frame.grid(row=0, column=0, columnspan=2, sticky="ew", pady=(0, 10))
         
-    def _setup_ui(self):
-        """Initialize the user interface."""
-        callbacks = {
-            'port_var': self.selected_port,
-            'autoscroll_var': self.autoscroll,
-            'toggle_connect': self.toggle_connection,
-            'toggle_run': self.toggle_running,
-            'clear_data': self.clear_data,
-            'export_excel': self.export_data,
-            'send_tare': self.send_tare_command
-        }
+        ttk.Label(conn_frame, text="Port:").grid(row=0, column=0, padx=(0, 5))
+        self.port_var = tk.StringVar()
+        self.port_combo = ttk.Combobox(conn_frame, textvariable=self.port_var, width=15)
+        self.port_combo.grid(row=0, column=1, padx=(0, 10))
         
-        self.main_window.build_ui(callbacks)
+        self.btn_refresh = ttk.Button(conn_frame, text="Refresh", command=self._refresh_ports)
+        self.btn_refresh.grid(row=0, column=2, padx=(0, 10))
         
-    def _start_periodic_tasks(self):
-        """Start periodic background tasks."""
-        self._schedule_port_refresh()
-        self._schedule_data_refresh()
+        self.btn_connect = ttk.Button(conn_frame, text="Connect", command=self._toggle_connection)
+        self.btn_connect.grid(row=0, column=3, padx=(0, 10))
         
-    # Connection Management
-    def toggle_connection(self):
-        """Toggle serial port connection."""
+        self.status_label = ttk.Label(conn_frame, text="Disconnected")
+        self.status_label.grid(row=0, column=4, padx=(10, 0))
+        
+        # Control frame
+        ctrl_frame = ttk.LabelFrame(main_frame, text="Control", padding="5")
+        ctrl_frame.grid(row=1, column=0, columnspan=2, sticky="ew", pady=(0, 10))
+        
+        self.btn_run = ttk.Button(ctrl_frame, text="Start", command=self._toggle_run, state=tk.DISABLED)
+        self.btn_run.grid(row=0, column=0, padx=(0, 10))
+        
+        self.btn_zero = ttk.Button(ctrl_frame, text="Zero Position", command=self._send_zero, state=tk.DISABLED)
+        self.btn_zero.grid(row=0, column=1, padx=(0, 10))
+        
+        self.btn_clear = ttk.Button(ctrl_frame, text="Clear Data", command=self._clear_data, state=tk.DISABLED)
+        self.btn_clear.grid(row=0, column=2, padx=(0, 10))
+        
+        self.btn_export = ttk.Button(ctrl_frame, text="Export Data", command=self._export_data, state=tk.DISABLED)
+        self.btn_export.grid(row=0, column=3, padx=(0, 10))
+        
+        # Data visualization frame
+        viz_frame = ttk.LabelFrame(main_frame, text="Real-time Data", padding="5")
+        viz_frame.grid(row=2, column=0, columnspan=2, sticky="nsew", pady=(0, 10))
+        main_frame.rowconfigure(2, weight=1)
+        
+        # Create plot and table
+        plot_frame = ttk.Frame(viz_frame)
+        plot_frame.pack(side="top", fill="both", expand=True, pady=(0, 10))
+        
+        table_frame = ttk.Frame(viz_frame)
+        table_frame.pack(side="bottom", fill="x")
+        
+        self.plot = EncoderPlot(plot_frame)
+        self.table = DataTable(table_frame)
+        
+        # Initialize port list
+        self._refresh_ports()
+    
+    def _refresh_ports(self):
+        """Refresh the list of available serial ports."""
+        ports = get_available_ports()
+        self.port_combo['values'] = ports
+        if ports and not self.port_var.get():
+            self.port_var.set(ports[0])
+    
+    def _toggle_connection(self):
+        """Toggle serial connection."""
         if not self.connection_state.get():
-            self._connect()
+            port = self.port_var.get()
+            if not port:
+                messagebox.showerror("Error", "Please select a port")
+                return
+            
+            # Create and start serial thread
+            self.serial_thread = SerialThread(port, line_callback=self._on_serial_line)
+            if self.serial_thread.connect():
+                self.serial_thread.start_reading()
+                self.connection_state.set(True)
+                self.btn_connect.config(text="Disconnect")
+                self.btn_run.config(state=tk.NORMAL)
+                self.btn_zero.config(state=tk.NORMAL)
+                self.status_label.config(text=f"Connected: {port}")
+            else:
+                messagebox.showerror("Error", f"Failed to connect to {port}")
+                self.serial_thread = None
         else:
-            self._disconnect()
-            
-    def _connect(self):
-        """Establish serial connection."""
-        port = self.selected_port.get()
-        if not port:
-            DialogHelper.show_warning("Port Selection", "Please select a COM port first")
-            return
-            
-        try:
-            self.serial_manager.start_connection(
-                port, 
-                DEFAULT_BAUD_RATE, 
-                self._handle_serial_data
-            )
-            
-            self.connection_state.set(True)
-            self.main_window.update_connection_state(True, port)
-            
-        except Exception as e:
-            DialogHelper.show_error("Connection Error", f"Failed to connect: {str(e)}")
-            
-    def _disconnect(self):
-        """Close serial connection."""
-        self.serial_manager.stop_connection()
-        self.connection_state.set(False)
-        self.running.set(False)
-        self.main_window.update_connection_state(False)
-        self.main_window.update_running_state(False)
-        
-    # Data Collection Management  
-    def toggle_running(self):
-        """Toggle data collection."""
+            # Disconnect
+            if self.serial_thread:
+                self.serial_thread.disconnect()
+                self.serial_thread = None
+            self.connection_state.set(False)
+            self.running.set(False)
+            self.btn_connect.config(text="Connect")
+            self.btn_run.config(text="Start", state=tk.DISABLED)
+            self.btn_zero.config(state=tk.DISABLED)
+            self.btn_clear.config(state=tk.DISABLED)
+            self.btn_export.config(state=tk.DISABLED)
+            self.status_label.config(text="Disconnected")
+    
+    def _toggle_run(self):
+        """Toggle data acquisition."""
         if not self.running.get():
-            self._start_data_collection()
+            self.running.set(True)
+            self.btn_run.config(text="Stop")
+            self.btn_clear.config(state=tk.NORMAL)
+            self.btn_export.config(state=tk.NORMAL)
+            self.status_label.config(text="Running")
         else:
-            self._stop_data_collection()
-            
-    def _start_data_collection(self):
-        """Start collecting data."""
-        # Auto-tare when starting
-        self.send_tare_command(auto=True)
-        
-        self.running.set(True)
-        self.main_window.update_running_state(True)
-        
-    def _stop_data_collection(self):
-        """Stop collecting data."""
-        self.running.set(False)
-        self.main_window.update_running_state(False)
-        
-    # Serial Data Handling
-    def _handle_serial_data(self, line: str):
-        """Process incoming serial data."""
-        if not self.running.get():
+            self.running.set(False)
+            self.btn_run.config(text="Start")
+            self.status_label.config(text="Connected")
+    
+    def _send_zero(self):
+        """Send zero position command to ESP32."""
+        if self.serial_thread:
+            if self.serial_thread.send_command("zero"):
+                self.status_label.config(text="Zero position sent")
+            else:
+                messagebox.showerror("Error", "Failed to send zero command")
+    
+    def _clear_data(self):
+        """Clear all captured data."""
+        if not messagebox.askyesno("Clear Data", "Clear all captured data?"):
             return
-            
-        # Parse the incoming line
-        pulse_count, force_value = self.data_parser.parse_line(line)
         
-        # Update force display
-        current_force = self.data_parser.get_current_force()
-        self.main_window.update_force_display(current_force)
-        
-        # Add sample to buffer if we got pulse data
-        if pulse_count is not None:
-            with self.data_mutex:
-                self.data_buffer.add(pulse_count, force_value)
-                
-    # Commands
-    def send_tare_command(self, auto: bool = False):
-        """Send TARE command to ESP32."""
-        if not self.connection_state.get():
-            return
-            
-        success = self.serial_manager.send_command("TARE")
-        
-        if success:
-            if not auto:
-                self.main_window.update_status("TARE command sent")
-        else:
-            if not auto:
-                DialogHelper.show_error("Command Error", "Failed to send TARE command")
-                
-    # Data Management
-    def clear_data(self):
-        """Clear all collected data."""
-        if not DialogHelper.ask_yes_no("Clear Data", "Clear all captured data?"):
-            return
-            
-        with self.data_mutex:
-            self.data_buffer.clear()
-            
-        self.main_window.clear_table()
-        self.main_window.clear_plot()
-        self.main_window.update_status("Data cleared")
-        
-    def export_data(self):
+        with self.mutex:
+            self.buffer.clear()
+        self.table.clear()
+        self.plot.update_plot(self.buffer)
+        self.status_label.config(text="Data cleared")
+    
+    def _export_data(self):
         """Export data to Excel file."""
-        with self.data_mutex:
-            samples = self.data_buffer.get_samples_copy()
-            
-        if not samples:
-            DialogHelper.show_info("Export", "No data to export.")
+        if not self.buffer.samples:
+            messagebox.showwarning("Export", "No data to export")
             return
-            
-        filename = DialogHelper.ask_save_filename(
-            DEFAULT_EXPORT_EXTENSION, 
-            [("Excel", "*.xlsx"), ("CSV", "*.csv")]
-        )
         
+        filename = filedialog.asksaveasfilename(
+            defaultextension=".xlsx",
+            filetypes=[("Excel files", "*.xlsx"), ("CSV files", "*.csv")]
+        )
         if not filename:
             return
-            
+        
         try:
-            if filename.lower().endswith('.xlsx'):
-                success = DataExporter.export_to_excel(samples, filename)
+            with self.mutex:
+                data = [
+                    {
+                        "Time (ms)": f"{sample.time_ms:.3f}",  # Αλλάζω από .1f σε .3f
+                        "Pos": sample.pos if sample.pos else "",
+                        "CPS": sample.cps if sample.cps else "",
+                        "RPM": sample.rpm if sample.rpm else ""
+                    }
+                    for sample in self.buffer.samples
+                ]
+            
+            df = pd.DataFrame(data)
+            if filename.endswith('.xlsx'):
+                df.to_excel(filename, index=False)
             else:
-                success = DataExporter.export_to_csv(samples, filename)
-                
-            if success:
-                summary = DataExporter.get_export_summary(samples)
-                DialogHelper.show_info(
-                    "Export Complete", 
-                    f"Exported {summary['count']} samples to {filename}"
-                )
-            else:
-                DialogHelper.show_error("Export Error", "Failed to export data")
-                
+                df.to_csv(filename, index=False)
+            
+            messagebox.showinfo("Export", f"Exported {len(data)} samples to {filename}")
+            self.status_label.config(text=f"Data exported ({len(data)} samples)")
         except Exception as e:
-            DialogHelper.show_error("Export Error", f"Export failed: {str(e)}")
-            
-    # Periodic Tasks
-    def _schedule_port_refresh(self):
-        """Schedule periodic port list refresh."""
-        self._refresh_port_list()
-        self.root.after(PORT_REFRESH_INTERVAL_MS, self._schedule_port_refresh)
+            messagebox.showerror("Export Error", f"Failed to export data: {e}")
+    
+    def _on_serial_line(self, line: str):
+        """Process incoming serial data line."""
+        if not self.running.get():
+            return
         
-    def _refresh_port_list(self):
-        """Refresh the list of available COM ports."""
+        # Store the raw ESP32 output line
+        line = line.strip()
+        if line:  # Only process non-empty lines
+            with self.mutex:
+                self.buffer.add(line)
+    
+    def _setup_periodic_tasks(self):
+        """Setup periodic GUI updates."""
+        def update_displays():
+            if self.running.get():
+                with self.mutex:
+                    self.table.update_table(self.buffer)
+                    self.plot.update_plot(self.buffer)
+            
+            # Schedule next update
+            self.root.after(10, update_displays)  # 10ms refresh rate (100 Hz)
+        
+        # Start periodic updates
+        self.root.after(10, update_displays)
+    
+    def run(self):
+        """Start the GUI application."""
         try:
-            current_ports = get_available_ports()
-            current_selection = self.selected_port.get()
-            
-            self.main_window.update_port_list(current_ports, current_selection)
-            
-            # Update selection if current port is no longer available
-            if current_selection not in current_ports:
-                new_selection = current_ports[0] if current_ports else ""
-                self.selected_port.set(new_selection)
-                
-        except Exception:
-            pass  # Ignore errors in background task
-            
-    def _schedule_data_refresh(self):
-        """Schedule periodic UI data refresh."""
-        self._update_display()
-        self.root.after(UI_REFRESH_MS, self._schedule_data_refresh)
-        
-    def _update_display(self):
-        """Update table and plot with new data."""
-        with self.data_mutex:
-            samples = self.data_buffer.get_samples_copy()
-            
-        if not samples:
-            return
-            
-        # Update table with new rows
-        self._update_table(samples)
-        
-        # Update plot
-        self._update_plot(samples)
-        
-    def _update_table(self, samples: list[Sample]):
-        """Update the data table with new samples."""
-        # Get current number of rows in table
-        current_rows = len(self.main_window.tree.get_children())
-        new_samples = samples[current_rows:]
-        
-        # Add new rows
-        last_item_id = None
-        for sample in new_samples:
-            force_str = f"{sample.force:.3f}" if sample.force is not None else ""
-            row_data = (
-                f"{sample.t * 1000:.1f}",  # Time in milliseconds
-                sample.pulses,
-                sample.delta,
-                force_str
-            )
-            last_item_id = self.main_window.add_table_row(row_data)
-            
-        # Auto-scroll if enabled
-        if self.autoscroll.get() and last_item_id:
-            self.main_window.scroll_to_row(last_item_id)
-            
-    def _update_plot(self, samples: list[Sample]):
-        """Update the plot with current data."""
-        if not samples:
-            return
-            
-        # Decimate data if too many points
-        total_samples = len(samples)
-        if total_samples > MAX_PLOT_POINTS * 2:
-            step = max(1, total_samples // DECIMATE_TARGET)
-            plot_samples = samples[::step]
-        else:
-            plot_samples = samples[-MAX_PLOT_POINTS:]
-            
-        if not plot_samples:
-            return
-            
-        # Extract data for plotting
-        times = [s.t for s in plot_samples]
-        pulses = [s.pulses for s in plot_samples]
-        
-        # Update plot
-        self.main_window.update_plot(times, pulses)
-        
-    # Application Lifecycle
-    def on_closing(self):
-        """Handle application closing."""
-        self.serial_manager.stop_connection()
-        self.root.destroy()
+            self.root.mainloop()
+        finally:
+            # Cleanup on exit
+            if self.serial_thread:
+                self.serial_thread.disconnect()
 
 
-def main():
-    """Main application entry point."""
-    # Create and configure root window
-    root = tk.Tk()
-    root.geometry("1000x700")  # Set reasonable default size
-    
-    # Create application
-    app = EncoderApplication(root)
-    
-    # Handle window closing
-    root.protocol("WM_DELETE_WINDOW", app.on_closing)
-    
-    # Start the GUI event loop
-    root.mainloop()
-
-
-if __name__ == '__main__':
-    main()
+if __name__ == "__main__":
+    app = EncoderGUI()
+    app.run()
