+++ python_client/serial_handler.py
@@ -1,125 +1,110 @@
 """
-Serial communication handler for the Encoder GUI.
-Manages connection to ESP32 and data parsing.
+Serial communication handler for ESP32 encoder data.
 """
-
 import threading
 import time
-from typing import Optional, Callable
 import serial
 import serial.tools.list_ports
+from typing import Callable, Optional, List
 
 
-class SerialReader(threading.Thread):
-    """Thread-safe serial port reader for ESP32 communication."""
+class SerialThread(threading.Thread):
+    """Thread for handling serial communication with ESP32."""
     
-    def __init__(self, port_getter: Callable[[], str], baud: int, 
-                 line_callback: Callable[[str], None], stop_event: threading.Event):
-        """
-        Initialize the serial reader.
-        
-        Args:
-            port_getter: Function that returns the current port name
-            baud: Baud rate for serial communication
-            line_callback: Function to call when a line is received
-            stop_event: Event to signal thread shutdown
-        """
+    def __init__(self, port: str, baudrate: int = 115200, 
+                 line_callback: Optional[Callable[[str], None]] = None):
         super().__init__(daemon=True)
-        self.port_getter = port_getter
-        self.baud = baud
+        self.port = port
+        self.baudrate = baudrate
         self.line_callback = line_callback
-        self.stop_event = stop_event
         self.ser: Optional[serial.Serial] = None
-
+        self.running = False
+        self.stop_event = threading.Event()
+    
+    def connect(self) -> bool:
+        """Establish serial connection."""
+        try:
+            self.ser = serial.Serial(
+                port=self.port,
+                baudrate=self.baudrate,
+                timeout=0.1,
+                write_timeout=1.0
+            )
+            return True
+        except Exception as e:
+            print(f"Serial connection failed: {e}")
+            return False
+    
+    def disconnect(self):
+        """Close serial connection."""
+        self.stop()
+        if self.ser and self.ser.is_open:
+            self.ser.close()
+        self.ser = None
+    
+    def start_reading(self):
+        """Start the serial reading thread."""
+        if not self.running:
+            self.running = True
+            self.stop_event.clear()
+            self.start()
+    
+    def stop(self):
+        """Stop the serial reading thread."""
+        self.running = False
+        self.stop_event.set()
+        if self.is_alive():
+            self.join(timeout=2.0)
+    
     def run(self):
         """Main thread loop for reading serial data."""
-        while not self.stop_event.is_set():
-            port = self.port_getter()
-            if not port:
-                time.sleep(0.5)
+        buffer = ""
+        while self.running and not self.stop_event.is_set():
+            if not self.ser or not self.ser.is_open:
+                time.sleep(0.1)
                 continue
-                
-            try:
-                # Open serial connection
-                self.ser = serial.Serial(port, self.baud, timeout=0.2)
-                self.ser.reset_input_buffer()
-                
-                # Read lines until stopped
-                while not self.stop_event.is_set():
-                    line = self.ser.readline().decode(errors='ignore').strip()
-                    if line:
-                        self.line_callback(line)
-                        
-            except serial.SerialException:
-                # Connection failed, retry after delay
-                time.sleep(0.5)
-            finally:
-                # Clean up serial connection
-                if self.ser:
-                    try:
-                        self.ser.close()
-                    except Exception:
-                        pass
-                    self.ser = None
-
-    def send_command(self, command: str) -> bool:
-        """
-        Send a command to the ESP32.
-        
-        Args:
-            command: Command string to send
             
-        Returns:
-            True if command was sent successfully, False otherwise
-        """
+            try:
+                if self.ser.in_waiting > 0:
+                    chunk = self.ser.read(self.ser.in_waiting).decode('utf-8', errors='ignore')
+                    buffer += chunk
+                    
+                    while '\n' in buffer:
+                        line, buffer = buffer.split('\n', 1)
+                        line = line.strip()
+                        if line and self.line_callback:
+                            self.line_callback(line)
+                else:
+                    time.sleep(0.01)
+            except Exception as e:
+                print(f"Serial read error: {e}")
+                time.sleep(0.1)
+    
+    def send_command(self, command: str) -> bool:
+        """Send a command to the ESP32."""
         if not self.ser or not self.ser.is_open:
             return False
-            
         try:
             self.ser.write(f"{command}\n".encode())
             self.ser.flush()
             return True
-        except Exception:
+        except Exception as e:
+            print(f"Serial write error: {e}")
             return False
 
 
-class SerialManager:
-    """Manages serial port connections and commands."""
-    
-    def __init__(self):
-        self.reader: Optional[SerialReader] = None
-        self.stop_event = threading.Event()
-        
-    def start_connection(self, port: str, baud: int, line_callback: Callable[[str], None]):
-        """Start serial connection with specified parameters."""
-        self.stop_connection()
-        
-        self.stop_event.clear()
-        self.reader = SerialReader(
-            lambda: port, 
-            baud, 
-            line_callback, 
-            self.stop_event
-        )
-        self.reader.start()
-        
-    def stop_connection(self):
-        """Stop the current serial connection."""
-        if self.reader:
-            self.stop_event.set()
-            self.reader = None
-            
-    def send_command(self, command: str) -> bool:
-        """Send a command through the active connection."""
-        if self.reader:
-            return self.reader.send_command(command)
-        return False
-        
-    def is_connected(self) -> bool:
-        """Check if serial connection is active."""
-        return self.reader is not None and not self.stop_event.is_set()
+def get_available_ports() -> List[str]:
+    """Get list of available serial ports."""
+    ports = serial.tools.list_ports.comports()
+    return [port.device for port in ports]
 
 
-def get_available_ports() -> list[str]:
-    """Get list of available COM ports."""
-    return [port.device for port in serial.tools.list_ports.comports()]
+def find_esp32_port() -> Optional[str]:
+    """Try to automatically detect ESP32 port."""
+    ports = serial.tools.list_ports.comports()
+    for port in ports:
+        # Look for common ESP32 identifiers
+        if any(keyword in port.description.lower() for keyword in 
+               ['esp32', 'silicon labs', 'cp210', 'ch340', 'usb serial']):
+            return port.device
+    return None
